+++ packages/astro/test/content-collections-render.test.js
@@ -142,23 +142,11 @@ describe('Content Collections - render()', () => {
 		it('getCollection should return new instances of the array to be mutated safely', async () => {
 			const app = await fixture.loadTestAdapterApp();
 
-			let request = new Request('http://example.com/');
-			let response = await app.render(request);
-			let html = await response.text();
-			let $ = cheerio.load(html);
-			const firstText = $('li').first().text();
-
-			request = new Request('http://example.com/sort-blog-collection');
-			response = await app.render(request);
-			html = await response.text();
-			$ = cheerio.load(html);
-			assert.notEqual($('li').first().text(), firstText);
-
-			request = new Request('http://example.com/');
-			response = await app.render(request);
-			html = await response.text();
-			$ = cheerio.load(html);
-			assert.equal($('li').first().text(), firstText);
+			const request = new Request('http://example.com/get-collection-equality');
+			const response = await app.render(request);
+			const html = await response.text();
+			const $ = cheerio.load(html);
+			assert.equal($('[data-are-equal]').first().text(), 'false');
 		});
 	});
 
@@ -217,6 +205,26 @@ describe('Content Collections - render()', () => {
 			const h2 = $('h2');
 			assert.equal(h2.length, 1);
 			assert.equal(h2.attr('data-components-export-applied'), 'true');
+		});
+
+		it('Stops collecting CSS when reaching a propagation stopping point', async () => {
+			let response = await fixture.fetch('/blog/5-big-news', { method: 'GET' });
+			assert.equal(response.status, 200);
+
+			let html = await response.text();
+			let $ = cheerio.load(html);
+
+			// Includes the red button styles used in the MDX blog post
+			assert.ok($('head > style').text().includes('background-color:red;'));
+
+			response = await fixture.fetch('/blog/about', { method: 'GET' });
+			assert.equal(response.status, 200);
+
+			html = await response.text();
+			$ = cheerio.load(html);
+
+			// Does not include the red button styles not used in this page
+			assert.equal($('head > style').text().includes('background-color:red;'), false);
 		});
 	});
 });